fix[rust]: determine supertype of structs
